Initial commit of joint code for Group OSCORE Prototype
This contains:
ACE
Group OSCORE (as part of Californium)
Joint applications (for the applications using both ACE and Group OSCORE)

ACE has been modified to use the Californium that contains Group OSCORE.
Including the version of COSE it contains.
The CBOR version ACE uses also had to be updated to match Californium.