Adding things for testing with Peter including GRP resource
Preparing for testing with Peter van der Stok using server
with public IP at v-pgroscore.sics.se

The server to use is TestOSCoreRSGroupOSCORE
For client use AltOSCoreClientGroupOSCORE

The client will use a hardcoded OSCORE context to connect,
skipping the token posting. The server is configured with
the same context and configured to consider that a token
has been posted from the sender ID of the client already.