Fixed issue on AS when retrieving OSCORE Sender ID
Previously this code only worked for the JUnit tests,
but due to the way the Sender ID was retrieved it did not
work when the AS and client were separate applications.

The fix was to provide the whole exchange to the CoapReq
class from where the Sender ID (locally the Recipient ID)
can be retrieved.