Fixed retrieval of Sender ID on AS to use parameters in endpoint context
Previously it used a different method that only worked locally
for the JUnit tests. This is since it used the CoAP token to
retrieve the OSCORE context and it relied on the client part
of the JUnit test to first send a request that would be stored
in the context database with that token as key.