Adapting the ACE code to the updated version of Californium
A lot of things changed in Californium since I updated to the latest
version, including things related to Scandium. In addition I also
needed to update the CBOR and Numbers versions. Due to this a lot
of things in the ACE code needs to be changed to adapt to this.
I am working on doing this.