Updated with Group OSCORE code from its branch in my Californium fork
This commit updates the Group OSCORE code to the latest version. It also
updates Californium code since I had rebased on the latest Californium
master in my fork.

The reason for deleting a lot of files in COSE is because ACE can now
use the COSE it has as a dependency and does not need to rely on the
COSE in Californium. Due to this, things not needed by Group OSCORE in
COSE could be deleted.

I have the Group OSCORE code under namespace groscore and COSE under
grcose so it will not conflict with the Californium code used as a
dependency by ACE.