Added the logic to react to a notification from the AS without waiting the next resource request.
Previously, after the Client obtained a resource, it slept for the entire duration of the request interval. So, it noticed that a token was revoked only after that time.

Now, the Client waits for the request interval, but if a token is removed from the valid tokens, it wakes up and promptly asks for a new token, if needed.

Also, the time at which the Client learns that a token was revoked is logged.