Changed DEFAULT_MAX_DENIALS value to Integer.MAX_INDEX.
As in our test we are not interested in Clients "assuming" token revocation, we set this variable to a high value, so we don't need to specify it as program argument each time we run a test.
The only test we use it is the one in which no entity uses the notification revocation mechanism. So, the RS uses introspection and the Client uses the "denials" (a number of 4.01 responses from the RS) to assume that the token eas revoked.