Added warm up. The Client can specify the -W option to request a Token with scope 'r_humidity' and audience 'rs2' when it is first initialized. After that, the performance at the Client are recorded.
On the other hand, tT, tS, and tE at the AS are recorded also for the warm-uo Token. So, even if -W is specified at the Client, tT1 (actually, t1T1 and t2T1) still refers to the first Token issued, i.e., the warm-up token.

The AS defines aa PIP for the 'r_humidity' attribute.

For the warm-up Token to be issued, the AS must be initialized with the correct scopes and audience for the Client, and with the RS rs2 managing the scope 'r_humidity'.

An example of AS arguments that correctly handles the warm-up Token is the following:

-F -R -n "rs1" -s "r_temp r_helloWorld" -u "rs1" -x "0x11" -m "RS1-AS-MS-------" -k "RS1-AS-Default-PSK-for-tokens---" -R -n "rs2" -s "r_humidity" -u "rs2" -x "0x12" -m "RS2-AS-MS-------" -k "RS2-AS-Default-PSK-for-tokens---" -C -n "ClientA" -s "r_temp r_helloWorld" -u rs1 -s "r_humidity" -u rs2 -x "0x22" -m "ClientA-AS-MS---"