Added -Y option to specify which resources the AS manages. This is used to initialize the related PIPs only if the resource is managed by the AS. For example, if "-Y Temp" is specified, a PIP monitoring the attribute "thermometer-reachable" will be initialized.
Moved the restoration of attribute values (those needed at the beginning of the run) in a method.

Removed PIPs and policies with multiple attributes (from 1 to n) for the helloWorld attribute.

Added a warmup policy that includes in the ongoing section a non-mutable attribute.
This is useful to issue a warmup token without having the UCS already monitoring some mutable attribute.
Moreover, with this solution, the request is enriched with only the mutable attributes used within the test.

The warmup policy is for the clientA, rs2, and r_warmup. So, if you want to use it, the AS must specify that in the CLI arguments. Something like this would work when specifying the resource server:
-R -n rs2 -s r_warmup -u rs2 -x 0x12 -m RS2-AS-MS------- -k RS2-AS-Default-PSK-for-tokens---

Also, the AS must define that clientA has the privileges:
-C -n ClientA -s r_warmup -u rs2 -x 0x22 -m ClientA-AS-MS---

A complete example of AS CLI arguments is the following:
-Y "Temp HelloWorld" -R -n rs1 -s "r_temp r_helloWorld" -u rs1 -x 0x11 -m RS1-AS-MS------- -k RS1-AS-Default-PSK-for-tokens--- -R -n rs2 -s r_warmup -u rs2 -x 0x12 -m RS2-AS-MS------- -k RS2-AS-Default-PSK-for-tokens--- -C -n ClientA -s "r_temp r_helloWorld" -u rs1 -s r_warmup -u rs2 -x 0x22 -m ClientA-AS-MS---

The client can access r_temp and r_helloWorld on rs1, and r_warmup on rs2.