Create a release instead of publishing to GitHub Packages
This is better because, to use the package that is created with this build, we don't need to authenticate to github.
In the pom.xml which wants to use this release as dependency, we will use the jitpack.io repository in order to access this jar without authenticating.
On the contrary, a jar published in the GitHub Packages requires authentication to be downloaded as dependency; this means that the .m2/settings.xml needs to be modified with the username and a token, and this is not what we want.

It seems that in the future, GitHub will allow public access to packages on GitHub Packages, but right now we have to use this workaround.