Change UcsHelper constructor
An additional parameter has been defined: policyTemplate.
This is supposed to be a string containing a basic xacml policy template, to be used when adding access for a client to a resource (addAccess method).
The addAccess method can also be called specifying a new, custom policy template, that will be used for adding the access.